+++ src/components/Pagination.tsx
@@ -1,25 +1,7 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import { colors } from '../styles';
-
-type PagiNationProps = {
-  totalPage: number;
-  limit: number;
-  page: number;
-  setPage: (page: number) => void;
-};
-
-const sliceArrayByLimit = (totalPage: number, limit: number) => {
-  const pages = [];
-  for (let i = 1; i <= totalPage; i++) {
-    pages.push(i);
-  }
-  const result = [];
-  for (let i = 0; i < pages.length; i += limit) {
-    result.push(pages.slice(i, i + limit));
-  }
-  return result;
-};
+import usePagination, { PagiNationProps } from '../hooks/usePagination';
 
 const Pagination: React.FC<PagiNationProps> = ({
   totalPage,
@@ -27,50 +9,27 @@ const Pagination: React.FC<PagiNationProps> = ({
   page,
   setPage,
 }) => {
-  const [currentPageArray, setCurrentPageArray] = useState<number[]>([]);
-  const [totalPageArray, setTotalPageArray] = useState<number[][]>([]);
+  const { currentPageArray } = usePagination({
+    totalPage,
+    limit,
+    page,
+    setPage,
+  });
 
-  // 컴포넌트 마운트 시 URL에서 페이지 번호를 가져와 상태 초기화
-  useEffect(() => {
-    const searchParams = new URLSearchParams(window.location.search);
-    const pageParam = parseInt(searchParams.get('page') || '1', 10);
-    if (pageParam >= 1 && pageParam <= totalPage) {
-      setPage(pageParam);
-    } else {
-      setPage(1); // 유효하지 않은 경우 기본값 설정
-    }
-  }, [totalPage, setPage]);
-
-  useEffect(() => {
-    const slicedPageArray = sliceArrayByLimit(totalPage, limit);
-    setTotalPageArray(slicedPageArray);
-    setCurrentPageArray(slicedPageArray[0] || []);
-  }, [totalPage, limit]);
-
-  useEffect(() => {
-    if (totalPageArray.length > 0) {
-      if (page % limit === 1) {
-        setCurrentPageArray(totalPageArray[Math.floor(page / limit)]);
-      } else if (page % limit === 0) {
-        setCurrentPageArray(totalPageArray[Math.floor(page / limit) - 1]);
-      }
-    }
-  }, [page, totalPageArray, limit]);
-
-  // 페이지 변경 시 URL 업데이트
-  useEffect(() => {
+  const updateURL = (newPage: number) => {
     const newSearchParams = new URLSearchParams(window.location.search);
-    newSearchParams.set('page', page.toString());
+    newSearchParams.set('page', newPage.toString());
     window.history.pushState(
       {},
       '',
       `${window.location.pathname}?${newSearchParams}`
     );
-  }, [page]);
+  };
 
   const handlePageChange = (newPage: number) => {
     if (newPage >= 1 && newPage <= totalPage) {
       setPage(newPage);
+      updateURL(newPage);
     }
   };
 
@@ -109,11 +68,8 @@ const S = {
   `,
   PageButton: styled.button<{ isActive: boolean }>`
     background-color: ${({ isActive }) =>
-      isActive
-        ? `${colors.semantic.primary}`
-        : `${colors.semantic.background.light}`};
-    color: ${({ isActive }) =>
-      isActive ? `${colors.semantic.text.light}` : ''};
+      isActive ? colors.semantic.primary : colors.semantic.background.light};
+    color: ${({ isActive }) => (isActive ? colors.semantic.text.light : '')};
     border-radius: 8px;
     padding: 0.26vw 0.625vw;
     cursor: pointer;
@@ -121,8 +77,8 @@ const S = {
     &:hover {
       background-color: ${({ isActive }) =>
         isActive
-          ? `${colors.semantic.hover.primary}`
-          : `${colors.semantic.hover.secondary}`};
+          ? colors.semantic.hover.primary
+          : colors.semantic.hover.secondary};
     }
     &:disabled {
       background-color: ${colors.semantic.disabled};
